♻️  Finish refactoring and add postman

+++ src/index.ts
@@ -9,22 +9,14 @@ import swaggerUi from 'swagger-ui-express';
 import { db as mysql } from './config/mysql.config';
 import { usersRouter } from './models/users/users.router';
 import { handleNotFound, handleError } from './middleware/error-handler';
-import morgan from 'morgan';
 import { specs } from './shared/swagger-specs';
-import { WebpackHotModule } from './shared/webpack-hot-module';
+import { WebpackHotModule } from './interface/webpack-hot-module';
 
 declare const module: WebpackHotModule;
 
 (async () => {
   dotenv.config();
-
-  if (!process.env.PORT) {
-    process.exit(1);
-  }
-
-  const PORT: number = parseInt(process.env.PORT as string, 10);
   const app = express();
-
 
   app.use(helmet());
   app.use((req, res, next) => {
@@ -35,10 +27,13 @@ declare const module: WebpackHotModule;
   app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
   app.use(handleNotFound);
   app.use(handleError);
-  app.use(morgan('dev'));
 
+  if (!process.env.PORT) {
+    process.exit(1);
+  }
   await mysql.instance.sync({ force: true });
 
+  const PORT: number = parseInt(process.env.PORT as string, 10);
   const server = app.listen(PORT, () => {
     console.log(`Listening on port ${PORT}`);
   });
